run tests via lst_transform
